requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
WTForms==3.1.1
python-dotenv==1.0.0
openpyxl==3.1.2
reportlab==4.0.7
pandas==2.1.4
python-dateutil==2.8.2
numpy==1.26.4
Pillow==11.3.0